requirements.txt
streamlit==1.32.2
transformers==4.39.2
fpdf==1.7.2
python-docx==1.1.0
torch==2.2.2
sentencepiece==0.2.0
